[Refactor]:removed redux, will use recoil. Added Popup screen upon which any popup can be shown
