+++ package.json
@@ -11,7 +11,6 @@
   "dependencies": {
     "@prisma/client": "5.18.0",
     "@radix-ui/react-dialog": "^1.1.1",
-    "@reduxjs/toolkit": "^2.2.7",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "lucide-react": "^0.446.0",
@@ -21,7 +20,6 @@
     "react": "^18",
     "react-dom": "^18",
     "react-icons": "^5.3.0",
-    "react-redux": "^9.1.2",
     "tailwind-merge": "^2.5.2",
     "tailwindcss-animate": "^1.0.7",
     "zod": "^3.23.8"
