--- src/app/store/slices/isMobileSlice.ts
@@ -1,22 +1,0 @@
-// import { createSlice } from "@reduxjs/toolkit";
-
-// interface isMobileState{
-//     isMobile: boolean
-// }
-
-// const initialState:isMobileState = {
-//     isMobile : window.innerWidth < 820 ? true : false
-// }
-
-// export const isMobileSlice = createSlice({
-//     name: "isMobile",
-//     initialState,
-//     reducers:{
-//         checkWindowSize : (state) =>{
-//             state.isMobile = window.innerWidth < 820 ? true : false;
-//         }
-//     }
-// })
-
-// export const {checkWindowSize} = isMobileSlice.actions;
-// export default isMobileSlice.reducer;
