--- src/app/store/store.ts
@@ -1,13 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit'
-// import isMobileSlice from './slices/isMobileSlice'
-
-export const store = configureStore({
-  reducer: {
-    // isMobile : isMobileSlice
-  },
-})
-
-// Infer the `RootState` and `AppDispatch` types from the store itself
-export type RootState = ReturnType<typeof store.getState>
-// Inferred type: {posts: PostsState, comments: CommentsState, users: UsersState}
-export type AppDispatch = typeof store.dispatch
